% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hipr.R
\name{cvhiqr}
\alias{cvhiqr}
\title{Cross-validation function for hiqr}
\usage{
cvhiqr(
  X,
  Z,
  K = 5,
  lambda = NULL,
  lambda.min = sqrt(log(max(ncol(X), ncol(Z)))/nrow(X)),
  nlambda = 50,
  rho = 1,
  err = 1e-05,
  maxIter = 1000
)
}
\arguments{
\item{X}{data matrix of dimension n*px.}

\item{Z}{data matrix of dimension n*pz.}

\item{K}{the number of folds. Default is 5.}

\item{lambda}{user supplied tuning parameter; Default is NULL and the program compute its own
sequence based on \code{nlambda}.}

\item{lambda.min}{smallest value for lambda, as a fraction of lambda.max which is available when lambda is NULL. 
Default is sqrt(log(p)/n).}

\item{nlambda}{the length of the tuning parameter sequence which is available when lambda is NULL. Default is 50.}

\item{rho}{step parameter for the ADMM. Default is 1.}

\item{err}{the precision used to stop the convergence. Default is 1e-5. 
Iterations stop when average absolute parameter change is less than \code{err}.}

\item{maxIter}{Maximum number of iterations. Default is 1000.}
}
\value{
A list with components
\item{Ome}{the estimated pz*px precision matrix.}
\item{cvlambda}{the chosen lambda by cross-validation.}
\item{lambda}{the used lambda list for cross-validation.}
\item{cvloss}{the empirical loss of cross-validation related to lambda.}
}
\description{
Cross-validation function for hiqr
}
