% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{projlasso}
\alias{projlasso}
\title{standard Lasso estimator of the projection matrix}
\usage{
projlasso(Z, covzx, lambda, err_abs = 10^(-4), maxIter = 1000L, rho = 1)
}
\arguments{
\item{\code{Z}, }{a \eqn{n*p_z} input data matrix}

\item{\code{covzx}, }{a \eqn{p_z*p_x} covariance matrix between \code{Z} and \code{X}.}

\item{\code{err_abs}, }{the absolute tolerance precision used to stop the convergence of ADMM.}

\item{\code{maxIter}, }{Maximum number of iterations. Default is 1000.}

\item{\code{lambda}, }{a scalar}

\item{\code{rho}, }{initial step parameter for ADMM.}
}
\value{
A list with components
\item{Omega}{a list of sparse \eqn{p_z*p_x} matrices corresponding to lambda.}
\item{lambda}{the used lambda for the solution path.}
\item{niter}{the number of each iterations for each number of lambda.}
}
\description{
ADMM algorithm for high dimensional projection matrix with lasso
\deqn{\argmin_W tr(\frac{1}{2}W^{\top}\Sigma_z W-\Sigma_{zx}^{\top}W)+\lambda ||W||_1.}
}
