\name{TestSQR-package}
\alias{TestSQR-package}
\alias{TestSQR}
\docType{package}
\title{
\packageTitle{TestSQR}
}
\description{
\packageDescription{TestSQR}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{TestSQR}

Maintainer: \packageMaintainer{TestSQR}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = TestSQR)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{TestSQR}}
%\section{Documentation Index}{\packageIndices{TestSQR}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
