% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testsqr.p}
\alias{testsqr.p}
\title{The projected version of testsqr}
\usage{
testsqr.p(
  y,
  X,
  Z,
  tau = 0.5,
  kernel = "Gaussian",
  h = max(0.05, 0.5 * sqrt(log(ncol(Z)))/nrow(X)^{
     0.3
 }),
  K = 5,
  lambda = NULL,
  lambda.min = sqrt(log(max(ncol(X), ncol(Z)))/nrow(X)),
  nlambda = 50,
  rho = 1,
  err = 1e-05,
  maxIter = 1000
)
}
\arguments{
\item{X}{data matrix of dimension n*px.}

\item{Z}{data matrix of dimension n*pz.}

\item{tau}{the quantile level with value between (0,1). The default is 0.5.}

\item{kernel}{A character string specifying the choice of kernel function., the default is "Gaussian".
Choices are "Gaussian", "uniform" and "logistic"}

\item{h}{The bandwidth parameter for kernel smoothing. The default is \eqn{\max\{0.05,0.5*(\log p_z/n)^{0.3}\}}}

\item{K}{the number of folds. Default is 5.}

\item{lambda}{user supplied tuning parameter; Default is NULL and the program compute its own
sequence based on \code{nlambda}.}

\item{lambda.min}{smallest value for lambda, as a fraction of lambda.max which is available when lambda is NULL. 
Default is sqrt(log(max(pz,px))/n).}

\item{nlambda}{the length of the tuning parameter sequence which is available when lambda is NULL. Default is 50.}

\item{rho}{step parameter for the ADMM. Default is 1.}

\item{err}{the precision used to stop the convergence. Default is 1e-5. 
Iterations stop when average absolute parameter change is less than \code{err}.}

\item{maxIter}{Maximum number of iterations. Default is 1000.}
}
\value{
A list with components
\item{tau}{The chosen quantile level}
\item{kernel}{The chosen kernel}
\item{statistic}{The value of the proposed U-statistic}
\item{standard.deviation}{The value of the standard deviation of the test statistic}
\item{statistic_scale}{The statistic finally used for the test following N(0,1)}
\item{pvalue}{p-value of the test statistic}
}
\description{
The projected version of testsqr
}
