% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testsqr}
\alias{testsqr}
\title{The test procedure for partial parameteres under the smoothed quantile regression}
\usage{
testsqr(
  y,
  X,
  Z,
  tau = 0.5,
  kernel = "Gaussian",
  h = max(0.05, 0.5 * (log(ncol(Z))/nrow(X))^{
     0.3
 })
)
}
\arguments{
\item{y}{response vector with length \eqn{n}.}

\item{X}{data matrix of dimension \eqn{n \times p_{x}}.}

\item{Z}{data matrix of dimension \eqn{n \times p_{z}}.}

\item{tau}{the quantile level with value between (0,1). The default is 0.5.}

\item{kernel}{A character string specifying the choice of kernel function., the default is "Gaussian".
Choices are "Gaussian", "uniform" and "logistic"}

\item{h}{The bandwidth parameter for kernel smoothing. The default is \eqn{\max\{0.05,0.5*(\log p_z/n)^{0.3}\}}}
}
\value{
A list with components
\item{tau}{The chosen quantile level}
\item{kernel}{The chosen kernel}
\item{statistic}{The value of the proposed U-statistic}
\item{standard.deviation}{The value of the standard deviation of the test statistic}
\item{statistic_scale}{The statistic finally used for the test following N(0,1)}
\item{pvalue}{p-value of the test statistic}
}
\description{
The test procedure for partial parameteres under the smoothed quantile regression
}
